requirements.txt
streamlit==1.40.2
PyYAML==6.0.1
mlxtend==0.23.3
scikit-learn==1.5.0
seaborn==0.13.2
matplotlib==3.9.0
numpy==1.26.4
xgboost==2.1.1
imbalanced-learn==0.12.4
pandas==2.2.2
